feat: :sparkles: "Marca Personal" section done, new pages animation and ux improvements
I create a "ScrollToTop" component to organize the href direction linkend in some buttons so when clicked u see the start of the page and not the end. Also and "Animation Page" component, so the transitions between pages is smoother and elegant. And finally I finish the main points in the "Marca Personal" component, added connection to Whatsapp and to "Servicios".

Pending to change the background texture so it looks better.