refactor: :art: Refactor the vlog section, added new covers, took off the border of the hero photo, changed the "Consultoría de Imagen Personal" and "Consultoría de Marca" work methodology grid,
In the vlog section I added a carrousel for the videos instead of the grid, linked the actual videos and added the actual covers, text, categories  etc. In services I refactored the work methodology grid of the 2 fist services. Minimum changes in the hero

Need to change the blue color of the dots in the bottom of the carrousel, refactor little details in the last 2 services and start working in the "Acerca de mí" section.