style: :construction: Improving details in several pages, so everything now is perfect, except the "Sobre mí" where I made changes, but I'm still working on it.
I linked the info in "ContactPreview" and "Contact", so now it looks so clean. I also took off the "AnimationPage" from "Servicios" and the subpages of it, also improve some details of the subpages, like text size and color, I also did the same with "El algoritmo de tu Closet". In Vlog I change some colors of the icons

I refactored the "Sobre mí" and made the web look faster without the AnimationPage thing.